requirements.txt
python-telegram-bot>=20.0
apscheduler>=3.10
aiohttp>=3.9
aiosqlite>=0.19
loguru>=0.7
python-dotenv>=1.0
py-clob-client>=0.1
matplotlib>=3.8
numpy>=1.24
openai>=1.0
